database: Move the searcher into separate file.
