+++ src/database/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod searcher;
+
 use fs_extra::dir;
 use r2d2::PooledConnection;
 use r2d2_sqlite::SqliteConnectionManager;
@@ -30,10 +32,11 @@ use zeroize::Zeroizing;
 use crate::config::{Config, LoadConfig, LoadDirection, SearchConfig};
 use crate::error::{Error, Result};
 use crate::events::{
-    CrawlerCheckpoint, Event, EventContext, EventId, HistoricEventsT, MxId, Profile,
+    CrawlerCheckpoint, Event, EventContext, EventId, HistoricEventsT, Profile,
     SerializedEvent,
 };
-use crate::index::{Index, IndexSearcher, Writer as IndexWriter};
+use crate::index::{Index, Writer as IndexWriter};
+pub use crate::database::searcher::{Searcher, SearchResult};
 
 #[cfg(test)]
 use fake::{Fake, Faker};
@@ -56,21 +59,6 @@ pub(crate) enum ThreadMessage {
     Write(Sender<Result<()>>, bool),
 }
 
-#[derive(Debug, PartialEq, Default, Clone)]
-/// A search result
-pub struct SearchResult {
-    /// The score that the full text search assigned to this event.
-    pub score: f32,
-    /// The serialized source of the event that matched a search.
-    pub event_source: SerializedEvent,
-    /// Events that happened before our matched event.
-    pub events_before: Vec<SerializedEvent>,
-    /// Events that happened after our matched event.
-    pub events_after: Vec<SerializedEvent>,
-    /// The profile of the sender of the matched event.
-    pub profile_info: HashMap<MxId, Profile>,
-}
-
 /// Statistical information about the database.
 pub struct DatabaseStats {
     /// The number number of bytes the database is using on disk.
@@ -79,12 +67,6 @@ pub struct DatabaseStats {
     pub event_count: u64,
     /// The number of rooms that the database knows about.
     pub room_count: u64,
-}
-
-/// The main entry point to the index and database.
-pub struct Searcher {
-    inner: IndexSearcher,
-    database: Arc<PooledConnection<SqliteConnectionManager>>,
 }
 
 struct Writer {
@@ -150,34 +132,6 @@ impl Writer {
         Ok(())
     }
 }
-
-impl Searcher {
-    /// Search the index and return events matching a search term.
-    /// # Arguments
-    ///
-    /// * `term` - The search term that should be used to search the index.
-    /// * `config` - A SearchConfig that will modify what the search result
-    /// should contain.
-    ///
-    /// Returns a list of `SearchResult`.
-    pub fn search(&self, term: &str, config: &SearchConfig) -> Result<Vec<SearchResult>> {
-        let search_result = self.inner.search(term, config)?;
-
-        if search_result.is_empty() {
-            return Ok(vec![]);
-        }
-
-        Ok(Database::load_events(
-            &self.database,
-            &search_result,
-            config.before_limit,
-            config.after_limit,
-            config.order_by_recency,
-        )?)
-    }
-}
-
-unsafe impl Send for Searcher {}
 
 /// The Seshat database.
 pub struct Database {
